# Conflicts:
#	src/Main.java
#	src/body/interfaces/Body.java
#	src/body/spaceship/SpaceShip.java
#	src/controllers/Controller.java
#	src/controllers/FeedBackController.java
#	src/controllers/PID.java
#	src/general_support/Vector.java
#	src/simulation/universe/Universe.java
#	src/simulation/universe/UniverseImpl.java

+++ src/simulation/Timer.java
@@ -1,13 +1,13 @@
 package simulation;
 
 public class Timer {
-    private long timePassedS = 0;
+    private double timePassedS = 0;
 
-    public void iterate(int timeStep) {
+    public void iterate(double timeStep) {
         timePassedS += timeStep;
     }
 
     long timePassedS() {
-        return timePassedS;
+        return (long) timePassedS;
     }
 }
